+++ src/del.js
@@ -11,7 +11,7 @@
  * // -> originalArray ['a', 'b', 'c', 'd', 'e']
  * // -> resultArray ['a', 'b', 'd', 'e']
  *
- * const resultArray2 = immutableDelete(originalArray, -1);
+ * const resultArray2 = immutableArrays.del(originalArray, -1);
  * // -> originalArray ['a', 'b', 'c', 'd', 'e']
  * // -> resultArray2 ['a', 'b', 'c', 'd', 'e']
  */
