+++ src/shift.js
