+++ src/splice.js
